requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0

# Deep Learning
tensorflow>=2.13.0
keras>=2.13.0

# Machine Learning
scikit-learn>=1.3.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Data processing
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.25.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0

